requirements.txt
pypdf2==3.0.1
chromadb==0.4.22
sentence-transformers==2.7.0
google-genai==0.3.0